+++ Projektskabelon-master/src/test/java/FunctionLayer/OverlayCalculatorTest.java
@@ -80,7 +80,7 @@ public class OverlayCalculatorTest {
         int numberOfFyr= OverlaySizeCalculator.numberOfFyrOnDistance(distance)*numberOfdistances;
 
         int lengthOfeight=fyrLengths.get(7);
-        int expectedLength=ConstructionSizeCalculator.raising(10,5400)+2000-360;
+        int expectedLength= (int)ConstructionSizeCalculator.raising(10,5400)+2000-360;
         assertEquals(fyrLengths.size(),numberOfFyr);
         assertEquals(expectedLength,lengthOfeight);
